Rename and move helper datastructs around
