+++ game/definitions/schematics.lua
