+++ game/domain/transformers/maze.lua
@@ -1,11 +1,8 @@
 
 -- dependencies
-local HELPERS = require 'lux.pack' 'domain.transformers.helpers'
-local Vector2  = require 'cpml.modules.vec2'
-
-local Rectangle  = HELPERS.rect
-local RANDOM     = HELPERS.random
-local SCHEMATICS = HELPERS.schematics
+local SCHEMATICS = require 'definitions.schematics'
+local RANDOM     = require 'common.random'
+local Vector2    = require 'cpml.modules.vec2'
 
 return function (_mapgrid, params)
   local _width, _height = _mapgrid.getDim()
@@ -157,5 +154,3 @@ return function (_mapgrid, params)
   end
   return caveMaze()
 end
-
-
