+++ game/tests/README.md
@@ -1,0 +1,13 @@
+
+# Tests
+
+## About
+
+Tests are important for us. When we develop a big new module, we want to make sure it is working by itself before shoving it into the other parts of the code.
+The files on this folder must be required from the `game/` folder. Why? Because that is where the game runs from! This is implemented so tests can run by themselves as well as while `ingame`.
+
+## List of Tests
+
+1. random
+2. map1
+3. bulkmap
