+++ game/tests/random.lua
@@ -1,7 +1,8 @@
 
-local random = require 'domain.transformers.helpers.random'
-local seed = tonumber(tostring(os.time()):sub(-7):reverse())
-math.randomseed(seed)
+local RANDOM = require 'common.random'
+local _seed = RANDOM.generateSeed()
+local setSeed = love and love.math.setRandomSeed or math.randomseed
+setSeed(seed)
 
 for i = 1, 16 do
   print("odd 0-20", random.odd(0, 20))
@@ -34,4 +35,3 @@ end
 for i = 1, 16 do
   print("even 1-19", random.even(1, 19))
 end
-
